// src/app/modules/academics/attendance/model/leave-request-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * LeaveRequestViewModel
 */
export interface LeaveRequestViewModel { 
    /**
     * Gets or Sets
     */
    studentId?: string;
    /**
     * Gets or Sets
     */
    reason?: string;
    /**
     * Gets or Sets FromDate
     */
    fromDate?: Date;
    /**
     * Gets or Sets ToDate
     */
    toDate?: Date;
    /**
     * Gets or Sets FileNamewithext
     */
    fileNamewithext?: string;
    /**
     * Gets or Sets FileValue
     */
    fileValue?: string;
}