// src/app/modules/academics/attendance/model/leave-status-change-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * LeaveStatusChangeViewMdel
 */
export interface LeaveStatusChangeViewModel { 
    /**
     * Gets or Sets LeaveId
     */
    leaveRequestId?: string;
    /**
     * Gets or Sets StatusId
     */
    statusId?: string;
}